refactor: SVG component common으로 이동
